- adjustments to model and cahce

+++ common/src/main/java/smartin/miapi/modules/properties/util/DoubleProperty.java
@@ -183,7 +183,7 @@ public abstract class DoubleProperty implements ModuleProperty {
 
     @Nullable
     public Double getValueRaw(ItemStack itemStack) {
-        return (Double) ModularItemCache.getRaw(itemStack, privateKey);
+        return ModularItemCache.getRaw(itemStack, privateKey);
     }
 
     public double getValueSafeRaw(ItemStack itemStack) {
